+++ bindings/qt/src/qzframe.cpp
@@ -149,19 +149,20 @@ void QZframe::setMore (int more)
 }
 
 ///
-//  Return frame routing id, set when reading frame from server socket
-//  or by the zframe_set_routing_id() method.                         
-size_t QZframe::routingId ()
+//  Return frame routing ID, if the frame came from a ZMQ_SERVER socket.
+//  Else returns zero.                                                  
+quint32 QZframe::routingId ()
 {
-    size_t rv = zframe_routing_id (self);
+    uint32_t rv = zframe_routing_id (self);
     return rv;
 }
 
 ///
-//  Set frame routing id. Only relevant when sending to server socket.
-void QZframe::setRoutingId (size_t routingId)
+//  Set routing ID on frame. This is used if/when the frame is sent to a
+//  ZMQ_SERVER socket.                                                  
+void QZframe::setRoutingId (quint32 routingId)
 {
-    zframe_set_routing_id (self, routingId);
+    zframe_set_routing_id (self, (uint32_t) routingId);
     
 }
